Handle the exception when storage file is missing
Code adapted from zyjames with modification

+++ src/seedu/addressbook/storage/StorageFile.java
@@ -19,6 +19,7 @@ import java.io.Reader;
 import java.io.Writer;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.io.File;
 
 /**
  * Represents the file used to store address book data.
@@ -27,7 +28,8 @@ public class StorageFile {
 
     /** Default file path used if the user doesn't provide the file name. */
     public static final String DEFAULT_STORAGE_FILEPATH = "addressbook.xml";
-
+	public static final String MESSAGE_STORAGE_FILE_NOT_FOUND = "Storage File could not be found, it might have been deleted";
+	
     /* Note: Note the use of nested classes below.
      * More info https://docs.oracle.com/javase/tutorial/java/javaOO/nested.html
      */
@@ -86,16 +88,28 @@ public class StorageFile {
         return filePath.toString().endsWith(".xml");
     }
 
+	/**
+     * Returns true if the file still exists.
+     */
+    private boolean isFileExist() {
+		File file = path.toFile();
+        return file.exists();
+    }
     /**
      * Saves all data to this storage file.
      *
      * @throws StorageOperationException if there were errors converting and/or storing data to file.
+	 * @throws FileNotFoundException if the file does not exist.
      */
-    public void save(AddressBook addressBook) throws StorageOperationException {
+    public void save(AddressBook addressBook) throws StorageOperationException, FileNotFoundException {
 
         /* Note: Note the 'try with resource' statement below.
          * More info: https://docs.oracle.com/javase/tutorial/essential/exceptions/tryResourceClose.html
          */
+		 
+		if (!isFileExist()) {
+			throw new FileNotFoundException(String.format(MESSAGE_STORAGE_FILE_NOT_FOUND, path));
+        }
         try (final Writer fileWriter =
                      new BufferedWriter(new FileWriter(path.toFile()))) {
 
@@ -115,8 +129,12 @@ public class StorageFile {
      * Loads data from this storage file.
      *
      * @throws StorageOperationException if there were errors reading and/or converting data from file.
+	 * @throws FileNotFoundException if the file does not exist.
      */
-    public AddressBook load() throws StorageOperationException {
+    public AddressBook load() throws StorageOperationException, FileNotFoundException {
+		if (!isFileExist()) {
+			throw new FileNotFoundException(String.format(MESSAGE_STORAGE_FILE_NOT_FOUND, path));
+        }
         try (final Reader fileReader =
                      new BufferedReader(new FileReader(path.toFile()))) {
